Massive redesign to avoid use of MaxUint64 as indicator of unknown size
Addresses #53
Not sure I like the changes, quite a lot of utilities assumed and worked
with uint64 integers. But this is no go if we want to use negative size
as indicator of an error or "do not use me" indicator.

Looks like we reducing size capacity by 2 (we are not using the sign bit).
I understand that if we using 2^63 file size, then we are in far future,
but I fill guilty by putting factor of 2 reduction.

What I like the least, I had replace uint64 with int64 in mio/stream

But if we OK with this, then changes pass all tests and ready to go.

+++ cmd/fs_handlers.go
@@ -226,7 +226,7 @@ func handleCp(ctx *cli.Context, ctl *client.Client) error {
 	return ctl.Copy(srcPath, dstPath)
 }
 
-func colorForSize(size uint64) func(f string, a ...interface{}) string {
+func colorForSize(size int64) func(f string, a ...interface{}) string {
 	switch {
 	case size >= 1024 && size < 1024<<10:
 		return color.CyanString
@@ -352,8 +352,8 @@ func handleList(ctx *cli.Context, ctl *client.Client) error {
 		fmt.Fprintf(
 			tabW,
 			"%s\t%s\t%s\t%s%s\t%s\t%s\n",
-			colorForSize(entry.Size)(humanize.Bytes(entry.Size)),
-			colorForSize(entry.Size)(humanize.Bytes(entry.CachedSize)),
+			colorForSize(entry.Size)(humanize.Bytes(uint64(entry.Size))),
+			colorForSize(entry.Size)(humanize.Bytes(uint64(entry.CachedSize))),
 			entry.ModTime.Format("2006-01-02 15:04:05 MST"),
 			userEntry,
 			coloredPath,
@@ -504,8 +504,8 @@ func handleShowFileOrDir(ctx *cli.Context, ctl *client.Client, path string) erro
 	printPair("Path", info.Path)
 	printPair("User", info.User)
 	printPair("Type", nodeType)
-	printPair("Size", fmt.Sprintf("%s (%d bytes)", humanize.Bytes(info.Size), info.Size))
-	printPair("Backend Size", fmt.Sprintf("%s (%d bytes)", humanize.Bytes(info.CachedSize), info.CachedSize))
+	printPair("Size", fmt.Sprintf("%s (%d bytes)", humanize.Bytes(uint64(info.Size)), info.Size))
+	printPair("Backend Size", fmt.Sprintf("%s (%d bytes)", humanize.Bytes(uint64(info.CachedSize)), info.CachedSize))
 	printPair("Inode", strconv.FormatUint(info.Inode, 10))
 	printPair("Pinned", pinState)
 	printPair("Explicit", explicitState)
